+++ dev/tools_docker.sh
@@ -10,7 +10,7 @@ ProjectDir=`pwd`
 
 # 設置環境
 Docker="king011/dev" # 發佈名稱
-DefaultTag=20.04 # 默認版本
+DefaultTag=24.04 # 默認版本
 Dockerfile="`pwd`/dockerfile" # 設置 dockerfile 檔案夾
 
 # 加載腳本