+++ flutter/CHANGELOG.md
@@ -1,3 +1,27 @@
+# 3.35.4
+
+| env         | version    |
+| ----------- | ---------- |
+| ubuntu      | 24.04      |
+| code-server | v4.104.2   |
+| flutter     | 3.35.4     |
+| openjdk     | 21         |
+| sdkmanager  | 12.0       |
+| build-tools | 34.0.0     |
+| platforms   | android-34 |
+
+# 3.27.4
+
+| env         | version    |
+| ----------- | ---------- |
+| ubuntu      | 22.04      |
+| code-server | v4.96.4    |
+| flutter     | 3.27.4     |
+| openjdk     | 18         |
+| sdkmanager  | 12.0       |
+| build-tools | 34.0.0     |
+| platforms   | android-34 |
+
 # 3.19.1
 
 | env         | version    |